feat: update landing page
