+++ apps/web/src/components/molecules/EventCard/AttendanceStatus.tsx
