+++ apps/web/src/components/molecules/EventCard/Thumbnail.tsx
